src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.16.0
openai==1.25.1
load-dotenv==0.1.0
python-dotenv==1.0.1